+++ pyhelios/wrappers/__init__.py
@@ -19,3 +19,4 @@ from . import UBoundaryLayerConductanceWrapper
 from . import UPhotosynthesisWrapper
 from . import UPlantArchitectureWrapper
 from . import ULeafOpticsWrapper
+from . import ULiDARWrapper